+++ lib/pages/tasques/edit_tasca.dart
@@ -255,12 +255,29 @@ class _LlistarCompraEditState extends State<LlistarCompraEdit> {
                       onPressed: () async {
                         final picked = await showDatePicker(
                           context: context,
+                          helpText: "SELECCIONA UNA DATA",
                           initialDate: model['dataPrevista'] == null
                               ? DateTime.now()
                               : DateTime.fromMicrosecondsSinceEpoch(
                                   model['dataPrevista'].microsecondsSinceEpoch),
                           firstDate: DateTime(2020),
                           lastDate: DateTime(2100),
+                          confirmText: "CONFIRMAR",
+                          cancelText: "CANCEL·LAR",
+                          builder: (BuildContext context, Widget child) {
+                            return Theme(
+                              data: ThemeData.dark().copyWith(
+                                colorScheme: ColorScheme.dark(
+                                  primary: Colors.orange[900],
+                                  onPrimary: Colors.white,
+                                  surface: Colors.orange,
+                                  onSurface: Colors.black87,
+                                ),
+                                dialogBackgroundColor: Colors.orange[50],
+                              ),
+                              child: child,
+                            );
+                          },
                         );
                         setState(() {
                           model['dataPrevista'] = picked == null
@@ -298,11 +315,24 @@ class _LlistarCompraEditState extends State<LlistarCompraEdit> {
                         final picked = await showDialog<int>(
                           context: context,
                           builder: (BuildContext context) {
-                            return new NumberPickerDialog.integer(
-                              title: Text("Temps estimat en h"),
-                              minValue: 1,
-                              maxValue: 100,
-                              initialIntegerValue: model['tempsEstimat'] ?? 1,
+                            return Theme(
+                              data: ThemeData.light().copyWith(
+                                accentColor: Colors.orange[900],
+                              ),
+                              child: NumberPickerDialog.integer(
+                                confirmWidget: Text(
+                                  "CONFIRMAR",
+                                  style: TextStyle(color: Colors.orange),
+                                ),
+                                cancelWidget: Text(
+                                  "CANCEL·LAR",
+                                  style: TextStyle(color: Colors.orange),
+                                ),
+                                title: Text("Temps estimat en h"),
+                                minValue: 1,
+                                maxValue: 100,
+                                initialIntegerValue: model['tempsEstimat'] ?? 1,
+                              ),
                             );
                           },
                         );